{{ cookiecutter.project_name }}/requirements.txt
flask_api==2.0
flask-marshmallow==0.12.0
Flask-RESTplus==0.13.0
flask-seasurf
Flask-SQLAlchemy==2.4.3
marshmallow-sqlalchemy==0.23
SQLAlchemy==1.3.17
redis==3.5.3
postgres==3.0.0
